implementación d3el crud

+++ src/main/java/com/erazo/springApp/models/dao/IClienteDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.erazo.springApp.models.entity.Cliente;
 
-public interface IClienteDao extends CrudRepository<Cliente, Integer>{
+public interface IClienteDao extends CrudRepository<Cliente, Long>{
 	
 }
